# services/auth/docs/swagger.yaml
basePath: /api/v1
definitions:
  app_auth-service_internal_common.ResponseAuth:
    properties:
      message:
        type: string
      session_jwt:
        type: string
    type: object
  app_auth-service_internal_model.Users:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      name:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      userUUID:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_user.RequestConfirm:
    properties:
      code:
        maximum: 999999
        minimum: 100000
        type: integer
      refresh_jwt:
        type: string
    required:
    - code
    type: object
  internal_user.RequestRemoveUser:
    properties:
      email:
        type: string
      password:
        maxLength: 24
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_user.RequestUpdateUser:
    properties:
      email:
        type: string
      new_email:
        type: string
      new_name:
        maxLength: 64
        minLength: 2
        type: string
      new_password:
        maxLength: 24
        minLength: 8
        type: string
      password:
        maxLength: 24
        minLength: 8
        type: string
    type: object
  internal_user.ResponseUser:
    properties:
      createdAt:
        type: string
      email:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      userUUID:
        type: string
    type: object
  response.NegativeResponse:
    properties:
      errors:
        type: object
      success:
        example: false
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      errors:
        additionalProperties:
          type: string
        type: object
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Core authentication and identity service with token rotation, profile
    management, and sliding window rate limiting.
  title: Auth & User Microservice API
  version: "1.0"
paths:
...
  /user:
    delete:
      consumes:
      - application/json
      description: Validates user credentials and the deletion type. If criteria are
        met, generates a confirmation session in Redis and triggers 2FA workflow (returns
        202 Accepted).
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Type of deletion: soft-delete or hard-delete'
        in: query
        name: type
        required: true
        type: string
      - description: Account credentials payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.RequestRemoveUser'
      produces:
      - application/json
      responses:
        "202":
          description: Deletion request accepted, 2FA confirmation workflow initiated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/app_auth-service_internal_common.ResponseAuth'
                errors:
                  type: object
              type: object
        "400":
          description: 'Client validation errors. Format: { \"errors\": { \"email\":
            \"incorrect email\" } } or { \"errors\": { \"action\": \"the type  must
            be a soft-delete or hard-delete\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "401":
          description: 'Authentication errors. Format: { \"errors\": { \"auth\": \"incorrect
            password or email\" } } or { \"errors\": { \"auth\": \"invalid access
            token\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "429":
          description: 'Too many requests. Format: { \"errors\": { \"global\": \"the
            limit for sending requests per minute has been exceeded\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "500":
          description: 'Server errors. Format: { \"errors\": { \"global\": \"critical
            error on the server side\" } } or { \"errors\": { \"global\": \"failed
            to remove user\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
      summary: Initiate account deletion workflow
      tags:
      - user
    get:
      description: Retrieves current authenticated user's profile details from MySQL
        by UUID stored in appjwt access token.
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile details successfully retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.ResponseUser'
                errors:
                  type: object
              type: object
        "401":
          description: 'Token errors. Format: { \"errors\": { \"auth\": \"invalid
            access token\" } } or { \"errors\": { \"auth\": \"access token has expired\"
            } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "404":
          description: 'Data errors. Format: { \"errors\": { \"user\": \"not found
            user\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "429":
          description: 'Too many requests. Format: { \"errors\": { \"global\": \"the
            limit for sending requests per minute has been exceeded\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "500":
          description: 'Server errors. Format: { \"errors\": { \"global\": \"critical
            error on the server side\" } } or { \"errors\": { \"global\": \"failed
            to get user\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
      summary: Get user profile
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Updates public info directly (returns 200 OK). If updating sensitive
        info (email/password), requires current password and triggers 2FA workflow
        (returns 202 Accepted).
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Profile update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.RequestUpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: Public profile details successfully updated instantly
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/app_auth-service_internal_model.Users'
                errors:
                  type: object
              type: object
        "202":
          description: Sensitive update accepted, 2FA confirmation workflow initiated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/app_auth-service_internal_common.ResponseAuth'
                errors:
                  type: object
              type: object
        "400":
          description: 'Validation or business logic errors. Format: { \"errors\":
            { \"new_name\": \"new name must be between 2 and 64 characters\" } } or
            { \"errors\": { \"new_password\": \"the new_password cannot  contain email\"
            } } or { \"errors\": { \"email\": \"either a new email or an old one must
            be present\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "401":
          description: 'Authentication errors. Format: { \"errors\": { \"auth\": \"incorrect
            password or email\" } } or { \"errors\": { \"auth\": \"invalid access
            token\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "404":
          description: 'Data errors. Format: { \"errors\": { \"user\": \"not found
            user\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "429":
          description: 'Too many requests. Format: { \"errors\": { \"global\": \"the
            limit for sending requests per minute has been exceeded\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "500":
          description: 'Server errors. Format: { \"errors\": { \"global\": \"critical
            error on the server side\" } } or { \"errors\": { \"global\": \"failed
            to update user\" } } or { \"errors\": { \"global\": \"failed to ensure
            security\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
      summary: Update user profile data or trigger sensitive update workflow
      tags:
      - user
  /user/confirm:
    post:
      consumes:
      - application/json
      description: Validates the 6-digit OTP code from Redis session. For updates,
        applies changes to MySQL and returns user data (200 OK). For hard/soft deletion,
        purges sessions, publishes to Kafka, and returns no content (204 No Content).
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Workflow action: update, soft-delete, or hard-delete'
        in: query
        name: action
        required: true
        type: string
      - description: Confirmation payload storing 6-digit code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.RequestConfirm'
      produces:
      - application/json
      responses:
        "200":
          description: Sensitive profile details successfully updated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.ResponseUser'
                errors:
                  type: object
              type: object
        "204":
          description: Account successfully deleted (soft or hard), no content returned
        "400":
          description: 'Client validation or workflow errors. Format: { \"errors\":
            { \"auth\": \"the code must be 6 characters\" } } or { \"errors\": { \"action\":
            \"the action must be update, soft-delete or hard-delete\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "401":
          description: 'Authentication or session errors. Format: { \"errors\": {
            \"auth\": \"invalid access token\" } } or { \"errors\": { \"auth\": \"refresh
            token has expired\" } } or { \"errors\": { \"auth\": \"incorrect refresh
            token\" } } or { \"errors\": { \"auth\": \"refresh token renewal error\"
            } } or { \"errors\": { \"auth\": \"incorrect session id\" } } or { \"errors\":
            { \"auth\": \"authorization session has expired or does not exist\" }
            } or { \"errors\": { \"auth\": \"incorrect code\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "404":
          description: 'Data errors. Format: { \"errors\": { \"user\": \"not found
            user\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "429":
          description: 'Too many requests. Format: { \"errors\": { \"global\": \"the
            limit for sending requests per minute has been exceeded\" } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
        "500":
          description: 'Server errors. Format: { \"errors\": { \"global\": \"failed
            to update user\" } } or { \"errors\": { \"global\": \"failed to delete
            user\" } } or { \"errors\": { \"global\": \"failed to remove user\" }
            } or { \"errors\": { \"global\": \"critical error on the server side\"
            } }'
          schema:
            $ref: '#/definitions/response.NegativeResponse'
      summary: Confirm 2FA code for profile updates or account deletion
      tags:
      - user
swagger: "2.0"
...